Fix error broken pipe by using a variable for the vms uuids instead of listing them using echo/printf.
Removed the /dev/null from error broken pipe which just redirects stderr to nothing.

Added extglob so uuids can be tested in one line using [[ instead of grep.

Since grep is external to the shell it is now removed.

The error 1 is replaced by 127 upon exit of systemd-vboxinit if it did not find VirtualBox or virtualbox.

This means that an executable is not found in your system. eg the exit status of cnf.

See http://en.wikipedia.org/wiki/Exit_status if you dont like man bash EXIT STATUS section.