Update to version 2.0.4
Remove the pkill in the ExtComm list.
Set 127 as the exit status if the ExtComms are not installed.
Change echo to printf in the MissingMessage.
Chage || to && and do not negate in the ExitMessage.
Declare the variables local to the specific functions.
Remove the echos.
Assign UpdateMessage in an array.
Add a daemonpid variable and savethe pid of all the daemons and avoid using pkill.
Align the indentations in the help and about messages.